feat(abstract-cosmos): sdk abstract-cosmos and sdk osmo refactored
Ticket: BG-79223
