+++ modules/abstract-cosmos/package.json
@@ -38,15 +38,16 @@
   },
   "dependencies": {
     "@bitgo/sdk-core": "^8.7.0",
-    "@bitgo/sdk-lib-mpc": "^8.4.0",
     "@bitgo/statics": "^16.1.0",
     "@bitgo/utxo-lib": "^9.0.0",
     "@cosmjs/amino": "^0.29.5",
+    "@cosmjs/crypto": "^0.30.1",
     "@cosmjs/encoding": "^0.29.5",
     "@cosmjs/proto-signing": "^0.29.5",
     "@cosmjs/stargate": "^0.29.5",
     "bignumber.js": "^9.1.1",
     "cosmjs-types": "^0.6.1",
+    "superagent": "^3.8.3",
     "lodash": "^4.17.21"
   },
   "devDependencies": {
