+++ modules/abstract-cosmos/src/cosmosCoin.ts
@@ -1,10 +1,8 @@
 import {
   BaseCoin,
+  BaseTransaction,
   BitGoBase,
-  Ecdsa,
-  ECDSA,
-  ECDSAMethodTypes,
-  hexToBigInt,
+  ExplanationResult,
   InvalidAddressError,
   InvalidMemoIdError,
   KeyPair,
@@ -12,22 +10,23 @@ import {
   ParsedTransaction,
   ParseTransactionOptions,
   SignedTransaction,
+  SigningError,
   SignTransactionOptions,
+  TransactionType,
   UnexpectedAddressError,
   VerifyAddressOptions,
   VerifyTransactionOptions,
 } from '@bitgo/sdk-core';
-import { EcdsaPaillierProof, EcdsaRangeProof, EcdsaTypes } from '@bitgo/sdk-lib-mpc';
-import { BaseCoin as StaticsBaseCoin, CoinFamily } from '@bitgo/statics';
+import { CoinFamily, BaseCoin as StaticsBaseCoin } from '@bitgo/statics';
 import { bip32 } from '@bitgo/utxo-lib';
 import { BigNumber } from 'bignumber.js';
 import { createHash, Hash, randomBytes } from 'crypto';
 import * as _ from 'lodash';
-import utils from './lib/utils';
-import * as url from 'url';
 import * as querystring from 'querystring';
+import * as request from 'superagent';
+import * as url from 'url';
 import { CosmosKeyPair } from './lib';
-import { Buffer } from 'buffer';
+import utils from './lib/utils';
 
 /**
  * Cosmos accounts support memo Id based addresses
@@ -58,6 +57,13 @@ export class CosmosCoin extends BaseCoin {
 
   static createInstance(bitgo: BitGoBase, staticsCoin?: Readonly<StaticsBaseCoin>): BaseCoin {
     return new CosmosCoin(bitgo, staticsCoin);
+  }
+
+  /**
+   * Creates an instance of TransactionBuilderFactory for the coin specific sdk
+   */
+  getBuilder(): any {
+    throw new Error('Method not implemented.');
   }
 
   /** @inheritDoc **/
@@ -100,16 +106,82 @@ export class CosmosCoin extends BaseCoin {
     return utils.isValidPrivateKey(prv);
   }
 
-  verifyTransaction(params: VerifyTransactionOptions): Promise<boolean> {
-    throw new Error('Method not implemented.');
-  }
-
   isValidAddress(address: string): boolean {
     throw new Error('Method not implemented.');
   }
 
-  signTransaction(params: SignTransactionOptions): Promise<SignedTransaction> {
-    throw new Error('Method not implemented.');
+  /** @inheritDoc **/
+  async verifyTransaction(params: VerifyTransactionOptions): Promise<boolean> {
+    let totalAmount = new BigNumber(0);
+    // const coinConfig = coins.get(this.getChain());
+    const { txPrebuild, txParams } = params;
+    const rawTx = txPrebuild.txHex;
+    if (!rawTx) {
+      throw new Error('missing required tx prebuild property txHex');
+    }
+    const transaction = await this.getBuilder().from(rawTx).build();
+    const explainedTx = transaction.explainTransaction();
+
+    if (txParams.recipients && txParams.recipients.length > 0) {
+      const filteredRecipients = txParams.recipients?.map((recipient) => _.pick(recipient, ['address', 'amount']));
+      const filteredOutputs = explainedTx.outputs.map((output) => _.pick(output, ['address', 'amount']));
+
+      if (!_.isEqual(filteredOutputs, filteredRecipients)) {
+        throw new Error('Tx outputs does not match with expected txParams recipients');
+      }
+      // WithdrawDelegatorRewards transaction doesn't have amount
+      if (transaction.type !== TransactionType.StakingWithdraw) {
+        for (const recipients of txParams.recipients) {
+          totalAmount = totalAmount.plus(recipients.amount);
+        }
+        if (!totalAmount.isEqualTo(explainedTx.outputAmount)) {
+          throw new Error('Tx total amount does not match with expected total amount field');
+        }
+      }
+    }
+    return true;
+  }
+
+  /** @inheritDoc **/
+  async explainTransaction(options: { txHex: string }): Promise<ExplanationResult> {
+    if (!options.txHex) {
+      throw new Error('missing required txHex parameter');
+    }
+    try {
+      const transactionBuilder = this.getBuilder().from(options.txHex);
+      const transaction = await transactionBuilder.build();
+      return transaction.explainTransaction();
+    } catch (e) {
+      throw new Error('Invalid transaction: ' + e.message);
+    }
+  }
+
+  /**
+   * Sign a transaction with a single private key
+   * @param params parameters in the form of { txPrebuild: {txHex}, prv }
+   * @returns signed transaction in the form of { txHex }
+   */
+  async signTransaction(
+    params: SignTransactionOptions & { txPrebuild: { txHex: string }; prv: string }
+  ): Promise<SignedTransaction> {
+    const txHex = params?.txPrebuild?.txHex;
+    const privateKey = params?.prv;
+    if (!txHex) {
+      throw new SigningError('missing required txPrebuild parameter: params.txPrebuild.txHex');
+    }
+    if (!privateKey) {
+      throw new SigningError('missing required prv parameter: params.prv');
+    }
+    const txBuilder = this.getBuilder().from(params.txPrebuild.txHex);
+    txBuilder.sign({ key: params.prv });
+    const transaction: BaseTransaction = await txBuilder.build();
+    if (!transaction) {
+      throw new SigningError('Failed to build signed transaction');
+    }
+    const serializedTx = transaction.toBroadcastFormat();
+    return {
+      txHex: serializedTx,
+    };
   }
 
   /** @inheritDoc **/
@@ -145,6 +217,91 @@ export class CosmosCoin extends BaseCoin {
     };
   }
 
+  /**
+   * Get the public node url from the Environments constant we have defined
+   */
+  protected getPublicNodeUrl(): string {
+    throw new Error('Method not implemented.');
+  }
+
+  /**
+   * Get account number from public node
+   */
+  protected async getAccountFromNode(senderAddress: string): Promise<request.Response> {
+    const nodeUrl = this.getPublicNodeUrl();
+    const getAccountPath = 'cosmos/auth/v1beta1/accounts/';
+    const fullEndpoint = nodeUrl + getAccountPath + senderAddress;
+    try {
+      return await request.get(fullEndpoint).send();
+    } catch (e) {
+      console.debug(e);
+    }
+    throw new Error(`Unable to call endpoint ${getAccountPath + senderAddress} from node: ${nodeUrl}`);
+  }
+
+  /**
+   * Get balance from public node
+   */
+  protected async getBalanceFromNode(senderAddress: string): Promise<request.Response> {
+    const nodeUrl = this.getPublicNodeUrl();
+    const getBalancePath = 'cosmos/bank/v1beta1/balances/';
+    const fullEndpoint = nodeUrl + getBalancePath + senderAddress;
+    try {
+      return await request.get(fullEndpoint).send();
+    } catch (e) {
+      console.debug(e);
+    }
+    throw new Error(`Unable to call endpoint ${getBalancePath + senderAddress} from node: ${nodeUrl}`);
+  }
+
+  /**
+   * Get chain id from public node
+   */
+  protected async getChainIdFromNode(): Promise<request.Response> {
+    const nodeUrl = this.getPublicNodeUrl();
+    const getLatestBlockPath = 'cosmos/base/tendermint/v1beta1/blocks/latest';
+    const fullEndpoint = nodeUrl + getLatestBlockPath;
+    try {
+      return await request.get(fullEndpoint).send();
+    } catch (e) {
+      console.debug(e);
+    }
+    throw new Error(`Unable to call endpoint ${getLatestBlockPath} from node: ${nodeUrl}`);
+  }
+
+  /**
+   * Helper to fetch account balance
+   */
+  protected async getAccountBalance(senderAddress: string): Promise<string> {
+    const response = await this.getBalanceFromNode(senderAddress);
+    if (response.status !== 200) {
+      throw new Error('Account not found');
+    }
+    return response.body.balances[0].amount;
+  }
+
+  /**
+   * Helper to fetch chainId
+   */
+  protected async getChainId(): Promise<string> {
+    const response = await this.getChainIdFromNode();
+    if (response.status !== 200) {
+      throw new Error('Account not found');
+    }
+    return response.body.block.header.chain_id;
+  }
+
+  /**
+   * Helper to fetch account number
+   */
+  protected async getAccountDetails(senderAddress: string): Promise<string[]> {
+    const response = await this.getAccountFromNode(senderAddress);
+    if (response.status !== 200) {
+      throw new Error('Account not found');
+    }
+    return [response.body.account.account_number, response.body.account.sequence];
+  }
+
   /** @inheritDoc **/
   generateKeyPair(seed?: Buffer): KeyPair {
     if (!seed) {
@@ -160,6 +317,11 @@ export class CosmosCoin extends BaseCoin {
     };
   }
 
+  /**
+   * Retrieves the address from a public key.
+   * @param {string} pubKey - The public key.
+   * @returns {string} The corresponding address.
+   */
   getAddressFromPublicKey(pubKey: string): string {
     return new CosmosKeyPair({ pub: pubKey }).getAddress();
   }
@@ -178,6 +340,10 @@ export class CosmosCoin extends BaseCoin {
     return true;
   }
 
+  /**
+   * Retrieves the SHA256 hash function.
+   * @returns {Hash} The SHA256 hash function.
+   */
   getHashFunction(): Hash {
     return createHash('sha256');
   }
@@ -246,159 +412,5 @@ export class CosmosCoin extends BaseCoin {
       return false;
     }
     return memoIdNumber.gte(0);
-  }
-
-  private getKeyCombinedFromTssKeyShares(
-    userPublicOrPrivateKeyShare: string,
-    backupPrivateOrPublicKeyShare: string,
-    walletPassphrase?: string
-  ): [ECDSAMethodTypes.KeyCombined, ECDSAMethodTypes.KeyCombined] {
-    let backupPrv;
-    let userPrv;
-    try {
-      backupPrv = this.bitgo.decrypt({
-        input: backupPrivateOrPublicKeyShare,
-        password: walletPassphrase,
-      });
-      userPrv = this.bitgo.decrypt({
-        input: userPublicOrPrivateKeyShare,
-        password: walletPassphrase,
-      });
-    } catch (e) {
-      throw new Error(`Error decrypting backup keychain: ${e.message}`);
-    }
-
-    const userSigningMaterial = JSON.parse(userPrv) as ECDSAMethodTypes.SigningMaterial;
-    const backupSigningMaterial = JSON.parse(backupPrv) as ECDSAMethodTypes.SigningMaterial;
-
-    if (!userSigningMaterial.backupNShare) {
-      throw new Error('Invalid user key - missing backupNShare');
-    }
-
-    if (!backupSigningMaterial.userNShare) {
-      throw new Error('Invalid backup key - missing userNShare');
-    }
-
-    const MPC = new Ecdsa();
-
-    const userKeyCombined = MPC.keyCombine(userSigningMaterial.pShare, [
-      userSigningMaterial.bitgoNShare,
-      userSigningMaterial.backupNShare,
-    ]);
-
-    const userSigningKeyDerived = MPC.keyDerive(
-      userSigningMaterial.pShare,
-      [userSigningMaterial.bitgoNShare, userSigningMaterial.backupNShare],
-      'm/0'
-    );
-
-    const userKeyDerivedCombined = {
-      xShare: userSigningKeyDerived.xShare,
-      yShares: userKeyCombined.yShares,
-    };
-
-    const backupKeyCombined = MPC.keyCombine(backupSigningMaterial.pShare, [
-      userSigningKeyDerived.nShares[2],
-      backupSigningMaterial.bitgoNShare,
-    ]);
-
-    if (
-      userKeyDerivedCombined.xShare.y !== backupKeyCombined.xShare.y ||
-      userKeyDerivedCombined.xShare.chaincode !== backupKeyCombined.xShare.chaincode
-    ) {
-      throw new Error('Common keychains do not match');
-    }
-
-    return [userKeyDerivedCombined, backupKeyCombined];
-  }
-
-  // TODO(BG-78714): Reduce code duplication between this and eth.ts
-  private async signRecoveryTSS(
-    userKeyCombined: ECDSA.KeyCombined,
-    backupKeyCombined: ECDSA.KeyCombined,
-    txHex: string,
-    {
-      rangeProofChallenge,
-    }: {
-      rangeProofChallenge?: EcdsaTypes.SerializedNtilde;
-    } = {}
-  ): Promise<ECDSAMethodTypes.Signature> {
-    const MPC = new Ecdsa();
-    const signerOneIndex = userKeyCombined.xShare.i;
-    const signerTwoIndex = backupKeyCombined.xShare.i;
-
-    // Since this is a user <> backup signing, we will reuse the same range proof challenge
-    rangeProofChallenge =
-      rangeProofChallenge ?? EcdsaTypes.serializeNtildeWithProofs(await EcdsaRangeProof.generateNtilde());
-
-    const userToBackupPaillierChallenge = await EcdsaPaillierProof.generateP(
-      hexToBigInt(userKeyCombined.yShares[signerTwoIndex].n)
-    );
-    const backupToUserPaillierChallenge = await EcdsaPaillierProof.generateP(
-      hexToBigInt(backupKeyCombined.yShares[signerOneIndex].n)
-    );
-
-    const userXShare = MPC.appendChallenge(
-      userKeyCombined.xShare,
-      rangeProofChallenge,
-      EcdsaTypes.serializePaillierChallenge({ p: userToBackupPaillierChallenge })
-    );
-    const userYShare = MPC.appendChallenge(
-      userKeyCombined.yShares[signerTwoIndex],
-      rangeProofChallenge,
-      EcdsaTypes.serializePaillierChallenge({ p: backupToUserPaillierChallenge })
-    );
-    const backupXShare = MPC.appendChallenge(
-      backupKeyCombined.xShare,
-      rangeProofChallenge,
-      EcdsaTypes.serializePaillierChallenge({ p: backupToUserPaillierChallenge })
-    );
-    const backupYShare = MPC.appendChallenge(
-      backupKeyCombined.yShares[signerOneIndex],
-      rangeProofChallenge,
-      EcdsaTypes.serializePaillierChallenge({ p: userToBackupPaillierChallenge })
-    );
-
-    const signShares: ECDSA.SignShareRT = await MPC.signShare(userXShare, userYShare);
-
-    const signConvertS21 = await MPC.signConvertStep1({
-      xShare: backupXShare,
-      yShare: backupYShare, // YShare corresponding to the other participant signerOne
-      kShare: signShares.kShare,
-    });
-    const signConvertS12 = await MPC.signConvertStep2({
-      aShare: signConvertS21.aShare,
-      wShare: signShares.wShare,
-    });
-    const signConvertS21_2 = await MPC.signConvertStep3({
-      muShare: signConvertS12.muShare,
-      bShare: signConvertS21.bShare,
-    });
-
-    const [signCombineOne, signCombineTwo] = [
-      MPC.signCombine({
-        gShare: signConvertS12.gShare,
-        signIndex: {
-          i: signConvertS12.muShare.i,
-          j: signConvertS12.muShare.j,
-        },
-      }),
-      MPC.signCombine({
-        gShare: signConvertS21_2.gShare,
-        signIndex: {
-          i: signConvertS21_2.signIndex.i,
-          j: signConvertS21_2.signIndex.j,
-        },
-      }),
-    ];
-
-    const MESSAGE = Buffer.from(txHex, 'hex');
-
-    const [signA, signB] = [
-      MPC.sign(MESSAGE, signCombineOne.oShare, signCombineTwo.dShare, createHash('sha256')),
-      MPC.sign(MESSAGE, signCombineTwo.oShare, signCombineOne.dShare, createHash('sha256')),
-    ];
-
-    return MPC.constructSignature([signA, signB]);
   }
 }
