+++ modules/abstract-cosmos/src/lib/index.ts
@@ -1,9 +1,12 @@
 import * as CosmosConstants from './constants';
 export * from './iface';
 
+export { StakingActivateBuilder } from './StakingActivateBuilder';
+export { StakingDeactivateBuilder } from './StakingDeactivateBuilder';
+export { StakingWithdrawRewardsBuilder } from './StakingWithdrawRewardsBuilder';
 export { CosmosKeyPair } from './keyPair';
 export { CosmosTransaction } from './transaction';
 export { CosmosTransactionBuilder } from './transactionBuilder';
-export { CosmosLikeUtils } from './utils';
-
+export { CosmosTransferBuilder } from './transferBuilder';
+export { CosmosUtils } from './utils';
 export { CosmosConstants };
