+++ modules/sdk-coin-osmo/package.json
@@ -44,14 +44,9 @@
     "@bitgo/sdk-core": "^8.7.0",
     "@bitgo/statics": "^16.1.0",
     "@cosmjs/amino": "^0.29.5",
-    "@cosmjs/crypto": "^0.29.5",
     "@cosmjs/encoding": "^0.29.5",
-    "@cosmjs/proto-signing": "^0.29.5",
     "@cosmjs/stargate": "^0.29.5",
-    "bignumber.js": "^9.1.1",
-    "cosmjs-types": "^0.6.1",
-    "lodash": "^4.17.21",
-    "superagent": "^3.8.3"
+    "bignumber.js": "^9.1.1"
   },
   "devDependencies": {
     "@bitgo/sdk-api": "^1.10.5",
