+++ modules/sdk-coin-osmo/test/unit/keyPair.ts
@@ -1,6 +1,6 @@
 import assert from 'assert';
 import should from 'should';
-import { OsmoKeyPair as KeyPair } from '../../src';
+import { KeyPair } from '../../src';
 
 import { TEST_ACCOUNT } from '../resources/osmo';
